ECAFE-2905 Доработка пакета синхронизации в части транзакций - не всегда загружалась сумма комплексов

+++ processor/src/core/java/ru/axetta/ecafe/processor/core/sync/response/AccountTransactionExtended.java
@@ -4,7 +4,6 @@
 
 package ru.axetta.ecafe.processor.core.sync.response;
 
-import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.Date;
 
@@ -79,7 +78,7 @@ public class AccountTransactionExtended {
         return complexsum;
     }
 
-    public void setComplexsum(BigDecimal complexsum) {
+    public void setComplexsum(BigInteger complexsum) {
         this.complexsum = complexsum.longValue();
     }
 
@@ -87,7 +86,7 @@ public class AccountTransactionExtended {
         return discountsum;
     }
 
-    public void setDiscountsum(BigDecimal discountsum) {
+    public void setDiscountsum(BigInteger discountsum) {
         this.discountsum = discountsum.longValue();
     }
